% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-kb.R
\name{cache_kb}
\alias{cache_kb}
\title{Keybase backed cache for cross-system caching}
\usage{
cache_kb(cache_path, algo = "sha512")
}
\arguments{
\item{cache_path}{Keybase filesystem sub-path directory for
storing cache files. Do not include \verb{K:} (Windows),
\verb{/keybase} (linux), \verb{/Volumes/keybase} (macOS) prefix.
Should start with one of \code{public/...}, \code{private/...}, or
\code{team/...}. See \url{https://book.keybase.io/docs/files} for more
information on how the Keybase filesystem works.}

\item{algo}{The hashing algorithm used for the cache, see \code{\link[digest:digest]{digest::digest()}}
for available algorithms.}
}
\description{
\href{https://keybase.io/}{Keybase} is a secure messaging and file-sharing
service that provides end-to-end encryption for all services. File-sharing
can be public, private (either to just you or to you and selected users),
or with a team. Files are synchronized across all systems where you
have logged in to Keybase. Presently, there is a 250GB storage limit.\cr
\cr
\strong{Keybase must be installed and the Keybase filesystem enabled to
use this package.}
}
\details{
This uses \code{\link[=cache_fs2]{cache_fs2()}}, which means two additional functions are
provided in slots in the returned list:
\itemize{
\item \code{location()} which will return the path to the cache directory
\item \code{size()} which will compute and return the size of the cache (in bytes)
}
}
