% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbmemoise-package.R
\docType{package}
\name{kbmemoise}
\alias{kbmemoise}
\alias{kbmemoise-package}
\title{Keybase Filesystem Backing Store for 'memoise' Caching}
\description{
Keybase \url{https://keybase.io} provides 250GB free, encrypted
storage making it suitable for creating a cross-system backing store
for 'memoise' caches. Tools are provided to use Keybase with 'memoise'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/kbmemoise}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/kbmemoise/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
