% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbmemoise-package.R
\docType{package}
\name{kbmemoise}
\alias{kbmemoise}
\alias{kbmemoise-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/kbmemoise}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/kbmemoise/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
